PythonWebAPI/requirement.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configparser==3.7.4
Flask==1.1.1
google-auth==1.6.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pyasn1==0.4.5
pyasn1-modules==0.2.5
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
